requirements.txt
Mesa==0.8.2
Flask_SQLAlchemy==2.3.2
Werkzeug==0.12.2
Flask_Login==0.4.0
Flask==0.12.2
setuptools==37.0.0
msgpack-python==0.4.8